Move SK.CoreSkills to new package: SK.Skills.Core
