+++ dotnet/src/Skills/Skills.UnitTests/Core/FileIOSkillTests.cs
@@ -4,10 +4,10 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.SemanticKernel;
-using Microsoft.SemanticKernel.CoreSkills;
+using Microsoft.SemanticKernel.Skills.Core;
 using Xunit;
 
-namespace SemanticKernel.UnitTests.CoreSkills;
+namespace SemanticKernel.Skills.UnitTests.Core;
 
 public class FileIOSkillTests
 {
