+++ dotnet/src/Skills/Skills.UnitTests/Core/TextSkillTests.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using Microsoft.SemanticKernel;
-using Microsoft.SemanticKernel.CoreSkills;
+using Microsoft.SemanticKernel.Skills.Core;
 using Xunit;
 
-namespace SemanticKernel.UnitTests.CoreSkills;
+namespace SemanticKernel.Skills.UnitTests.Core;
 
 public class TextSkillTests
 {
