+++ dotnet/src/Skills/Skills.UnitTests/Core/WaitSkillTests.cs
@@ -2,11 +2,12 @@
 
 using System.Threading.Tasks;
 using Microsoft.SemanticKernel;
-using Microsoft.SemanticKernel.CoreSkills;
+using Microsoft.SemanticKernel.Skills.Core;
 using Moq;
+using SemanticKernel.UnitTests;
 using Xunit;
 
-namespace SemanticKernel.UnitTests.CoreSkills;
+namespace SemanticKernel.Skills.UnitTests.Core;
 
 // TODO: allow clock injection and test all functions
 public class WaitSkillTests
