+++ samples/dotnet/kernel-syntax-examples/Example01_NativeFunctions.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System;
-using Microsoft.SemanticKernel.CoreSkills;
+using Microsoft.SemanticKernel.Skills.Core;
 
 // ReSharper disable once InconsistentNaming
 public static class Example01_NativeFunctions
